refactor(services): split publish, registration and rescore
